simulate/genesis/requirements.txt
genesis-world>=0.3.1
torch
skrl[torch]
pyyaml
hidapi
